Testing

+++ routes/userListings.js
@@ -5,9 +5,9 @@ const Listings = require("../models/listings");
 
 router.get("/:id",auth,  async (req, res) => {
   console.log('lets go')
-  const _id = req.params.id;
+  const userId = req.params.id;
  
-    const userListings = await Listings.find({ _id })
+    const userListings = await Listings.find({ userId })
 
   if (userListings.length < 1) return res.status(404).send([])
     console.log('userListings21')
@@ -19,10 +19,17 @@ router.get("/:id",auth,  async (req, res) => {
 
 router.delete("/:id",auth,  async (req, res) => {
   console.log('lets go')
-  const _id = req.params.id;
+  const userId = req.params.id;
+  const _id = req.params._id;
  
-    const userListings = await Listings.deleteOne({ _id })
-res.status(200).send(userListings);
+    const userListings = await Listings.deleteOne({ userId, _id })
+
+  if (userListings.length < 1) return res.status(404).send([])
+    console.log('userListings21')
+  console.log(userListings)
+  console.log(userId)
+  console.log('you say waiting')
+  res.status(200).send(userListings);
 });
 
 module.exports = router;
